Se corrigen errores: - Para la entity "Cliente" se cambio el tipado del atributo "celular" de Integer a Long - Para la entity "ComprasProducto" se corrigen las anotaciones @JoinColumn, ya que tenian el atributo 'referencedColumnName' y se cambia a 'name'

+++ src/main/java/co/com/marimaro/store/persistance/entity/Cliente.java
@@ -15,7 +15,7 @@ public class Cliente {
     private String id;
     private String nombre;
     private String apellidos;
-    private Integer celular;
+    private Long celular;
     private String direccion;
     @Column(name = "correo_electronico")
     private String correoEletronico;
